Fix bad type in near() signature

+++ test/lib/matchers.h
@@ -2,7 +2,7 @@
 
 #include "lib/catch.hpp"
 
-inline Catch::Matchers::Floating::WithinUlpsMatcher near(float expected, float proximity = 1) {
+inline Catch::Matchers::Floating::WithinUlpsMatcher near(float expected, int proximity = 1) {
   return Catch::Matchers::WithinULP(expected, proximity);
 }
